requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
strawberry-graphql
sqlalchemy
psycopg2-binary
